pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spare-paw"
version = "0.1.0"
description = "24/7 personal AI agent on Android via Termux + Telegram"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9",
    "apscheduler>=3.10,<4",
    "beautifulsoup4>=4.12",
    "tiktoken>=0.7",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "mcp>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
where = ["src"]
